+++ core/src/net/oschina/j2cache/J2Cache.java
@@ -22,8 +22,6 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 /**
  * J2Cache 的缓存入口
@@ -43,22 +41,18 @@ public class J2Cache {
 			initFromConfig();
 			/* 初始化缓存接口 */
 			channel = new CacheChannel(){
-				//为了避免发送广播的堵塞或者延迟导致的应用响应速度慢，因此采用线程方式发送
-				ExecutorService threadPool = Executors.newCachedThreadPool();
-
 				@Override
 				public void sendClearCmd(String region) {
-					threadPool.execute(()->policy.sendClearCmd(region));
+					policy.sendClearCmd(region);
 				}
 
 				@Override
 				public void sendEvictCmd(String region, String...keys) {
-					threadPool.execute(()->policy.sendEvictCmd(region, keys));
+					policy.sendEvictCmd(region, keys);
 				}
 
 				@Override
 				public void close() {
-					threadPool.shutdownNow();
 					policy.disconnect();
 					CacheProviderHolder.shutdown();
 				}