+++ core/src/net/oschina/j2cache/redis/RedisCacheProvider.java
@@ -49,10 +49,6 @@ public class RedisCacheProvider implements CacheProvider {
         return CacheObject.LEVEL_2;
     }
 
-    public RedisClient getClient() {
-        return redisClient;
-    }
-
     /**
      * 初始化 Redis 连接
      * @param props current configuration settings.
@@ -62,7 +58,7 @@ public class RedisCacheProvider implements CacheProvider {
         this.namespace = props.getProperty("namespace");
         this.storage = props.getProperty("storage");
 
-        JedisPoolConfig poolConfig = newPoolConfig(props);
+        JedisPoolConfig poolConfig = RedisUtils.newPoolConfig(props, null);
 
         String hosts = props.getProperty("hosts");
         String mode = props.getProperty("mode");
@@ -110,29 +106,6 @@ public class RedisCacheProvider implements CacheProvider {
     @Override
     public Cache buildCache(String region, long timeToLiveInSeconds, CacheExpiredListener listener) {
         return buildCache(region, listener);
-    }
-
-    /**
-     * 初始化 Redis 连接池
-     * @param props
-     * @return
-     */
-    private JedisPoolConfig newPoolConfig(Properties props) {
-        JedisPoolConfig cfg = new JedisPoolConfig();
-        cfg.setMaxTotal(Integer.valueOf((String)props.getOrDefault("maxTotal", "-1")));
-        cfg.setMaxIdle(Integer.valueOf((String)props.getOrDefault("maxIdle", "100")));
-        cfg.setMaxWaitMillis(Integer.valueOf((String)props.getOrDefault("maxWaitMillis", 100)));
-        cfg.setMinEvictableIdleTimeMillis(Integer.valueOf((String)props.getOrDefault("minEvictableIdleTimeMillis", "864000000")));
-        cfg.setMinIdle(Integer.valueOf((String)props.getOrDefault("minIdle", "10")));
-        cfg.setNumTestsPerEvictionRun(Integer.valueOf((String)props.getOrDefault("numTestsPerEvictionRun", "10")));
-        cfg.setLifo(Boolean.valueOf(props.getProperty("lifo", "false")));
-        cfg.setSoftMinEvictableIdleTimeMillis(Integer.valueOf((String)props.getOrDefault("softMinEvictableIdleTimeMillis", "10")));
-        cfg.setTestOnBorrow(Boolean.valueOf(props.getProperty("testOnBorrow", "true")));
-        cfg.setTestOnReturn(Boolean.valueOf(props.getProperty("testOnReturn", "false")));
-        cfg.setTestWhileIdle(Boolean.valueOf(props.getProperty("testWhileIdle", "false")));
-        cfg.setTimeBetweenEvictionRunsMillis(Integer.valueOf((String)props.getOrDefault("timeBetweenEvictionRunsMillis", "300000")));
-        cfg.setBlockWhenExhausted(Boolean.valueOf(props.getProperty("blockWhenExhausted", "true")));
-        return cfg;
     }
 
 }